+++ changelog.md
@@ -36,6 +36,7 @@ Codion Change Log
 - EntitySearchModel.Builder.searchColumns() renamed search(columns).
 - EntitySearchModel.Search.result() renamed perform().
 - EntityPersistence.replaceable() removed, unused.
+- DefaultEntityEditor, minor refactoring.
 ### is.codion.swing.framework.ui
 - EditorComponents.entries() and editorComponent() bug fixed, now exclude entries without a component.
 - EditorComponents.detail() added, removal implemented.